change from secret to config.yml with example.
add black as dev dependency.
refactor src structure.
remove python 2 support.
