+++ .gitignore
@@ -131,6 +131,3 @@ dmypy.json
 # Editor config
 .idea
 .vscode
-
-# config
-secret*
