+++ Pipfile
@@ -3,13 +3,23 @@ name = "pypi"
 url = "https://pypi.org/simple"
 verify_ssl = true
 
+[scripts]
+ui = "python src/ui.py"
+tests = "python -m unittest discover tests"
+lint = "black ."
+
 [dev-packages]
+black = "*"
 
 [packages]
 lxml = "==3.7.3"
 olefile = "==0.44"
 python-docx = "==0.8.6"
 Pillow = "==4.0.0"
+pyyaml = "*"
 
 [requires]
 python_version = "3.6"
+
+[pipenv]
+allow_prereleases = true
