+++ config.yml
@@ -1,0 +1,1 @@
+email: 'test@example.com' # change me (used in feedback)
